refactor(library/Attempt): moves "/Error/NonActionableBuiltInError.ts" to "/Error/NonActionableBuiltInError/definition.ts"
